refactor(hub): zero-config defaults (memory-store, sync-policy) back to the spine
memory-store.ts and sync-policy.ts aren't part of the to-* family
contract (port/to/index.ts never re-exported them) — they're the
kernel's own zero-config defaults, only reachable via createNoydb()'s
defaults and the root barrel. Move them from src/port/to/ back to
src/kernel/ alongside the orchestration files that actually use them,
and repoint every importer (noydb.ts, types.ts, src/index.ts, the
with-party/team sync files, and four test files).

Drop the now-obsolete PRE_EXISTING_SPINE_SERVICE_IMPORTS grandfather
entries for noydb.ts's and types.ts's memory-store/sync-policy
specifiers — after the move these are spine-internal imports, not
spine→port, so port-layering passes clean without them.
