+++ packages/hub/__tests__/sealed-query-refusal.test-d.ts
@@ -4,7 +4,7 @@
  */
 import { describe, it, expectTypeOf } from 'vitest'
 import { createNoydb, sum, money } from '../src/index.js'
-import { memoryStore } from '../src/port/to/memory-store.js'
+import { memoryStore } from '../src/kernel/memory-store.js'
 
 interface Person { id: string; name: string; ssn: string; age: number }
