+++ packages/hub/src/kernel/noydb.ts
@@ -88,8 +88,8 @@ import { NO_TX, type TxStrategy } from '../with-commit/tx/strategy.js'
 import { NO_FORGET, type ForgetStrategy } from '../with-audit/forget/strategy.js'
 import { NO_CUSTODY, type CustodyStrategy, type CustodyHost } from '../with-party/custody/strategy.js'
 import { readDottedPath, coerceSubjectId } from '../with-audit/forget/subject-index.js'
-import { INDEXED_STORE_POLICY } from '../port/to/sync-policy.js'
-import { memoryStore } from '../port/to/memory-store.js'
+import { INDEXED_STORE_POLICY } from './sync-policy.js'
+import { memoryStore } from './memory-store.js'
 import type { PolicyEnforcer } from '../with-party/session/session-policy.js'
 import { NO_SESSION, type SessionStrategy } from '../with-party/session/strategy.js'
 import {
